+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2019-12-13
+### Added
+- added additional Comprehend and Transcribe languages support to Media2Cloud
+
+### Changed
+- workaround audio analysis issue when Comprehend result contains leading punctuation characters.
+
+### Removed
+
 ## [2.0.0] - 2019-10-17
 ### Added
 - Backend: new analysis engine with analysis-monitor, audio-analysis, video-analysis, image-analysis state machines